* WIP Refactoring the report info gathering file * WIP Refactoring the report map information
