+++ monkey_island/cc/app.py
@@ -19,7 +19,7 @@ from cc.resources.monkey_configuration import MonkeyConfiguration
 from cc.resources.monkey_download import MonkeyDownload
 from cc.resources.netmap import NetMap
 from cc.resources.pthmap import PthMap
-from cc.resources.pthreporthtml import PthReportHtml
+from cc.resources.pthreport import Report
 from cc.resources.node import Node
 from cc.resources.report import Report
 from cc.resources.root import Root
@@ -107,6 +107,6 @@ def init_app(mongo_url):
     api.add_resource(TelemetryFeed, '/api/telemetry-feed', '/api/telemetry-feed/')
     api.add_resource(Log, '/api/log', '/api/log/')
     api.add_resource(PthMap, '/api/pthmap', '/api/pthmap/')
-    api.add_resource(PthReportHtml, '/api/pthreport', '/api/pthreport/')
+    api.add_resource(Report, '/api/pthreport', '/api/pthreport/')
 
     return app
